chore(services): rename rpc/services → rpc/handlers (PR 10/N)
The symbolic deliverable from the plan's "close the loop" PR. Renames
`apps/server/src/routes/rpc/services/` → `apps/server/src/routes/rpc/handlers/`
so the distinction between "the services layer" (owns business logic,
lives in `packages/services`) and "Connect transport handlers" (thin
proto → service → proto wrappers) is permanent and visible in the path.

Keeping the old name invites the next developer to "just add one small
thing" to a file under a `services/` folder months later; the rename
makes the layering explicit.

### Changes

- `git mv` of the six domain subdirectories + their tests
  (health / maintenance / monitor / notification / status-page /
  status-report).
- `router.ts` import paths updated from `./services/*` to `./handlers/*`.
- Biome `overrides.include` paths updated to the new location.
- Added `apps/server/src/routes/rpc/handlers/health/**` to the scope —
  the health handler has no db usage today; including it locks in that
  invariant.

### Still out of scope (follow-ups)

Rather than pretending the full "close the loop" deliverable is possible
today, the biome.jsonc comment now enumerates exactly what remains
unmigrated:

  - `packages/api/src/router/statusPage.ts` — public viewer endpoints
    under `publicProcedure`, no authed `ServiceContext`.
  - `packages/api/src/router/{member,integration,monitorTag,
    pageSubscriber,privateLocation,checker,feedback,stripe,tinybird,
    email}.ts` — small domains not yet lifted.
  - `apps/server/src/routes/rpc/handlers/monitor/**` — 6 jobType-specific
    methods still on db.
  - `apps/server/src/routes/rpc/handlers/status-page/**` — page CRUD is
    migrated (PR 7), but components / groups / subscribers / viewer (13
    methods) still import db, so the whole file stays out of scope.
  - `apps/server/src/routes/v1/**` — the public HTTP API surface.
  - `apps/server/src/routes/slack/**` except `interactions.ts` — tools,
    handler, oauth, workspace-resolver still on db.
  - `apps/server/src/routes/public/**` — public-facing HTTP routes.

Each of the above is its own PR-sized migration. The final consolidation
(broadening to `router/**` + dropping `@openstatus/db` from
`packages/api` and `apps/server`) is conditional on all of them
landing first.
