+++ apps/server/src/routes/rpc/handlers/health/index.ts
