+++ apps/server/src/routes/rpc/handlers/maintenance/errors.ts
