+++ apps/server/src/routes/rpc/handlers/maintenance/index.ts
