+++ apps/server/src/routes/rpc/handlers/monitor/__tests__/monitor.test.ts
