+++ apps/server/src/routes/rpc/handlers/monitor/converters/enums.ts
