+++ apps/server/src/routes/rpc/handlers/monitor/errors.ts
