+++ apps/server/src/routes/rpc/handlers/monitor/index.ts
