+++ apps/server/src/routes/rpc/handlers/notification/errors.ts
