+++ apps/server/src/routes/rpc/handlers/status-page/errors.ts
