+++ apps/server/src/routes/rpc/handlers/status-page/index.ts
